+++ boreas/boreas_io.h
@@ -42,10 +42,10 @@ void
 send_dead_hosts_to_ospd_openvas (int);
 
 void
-init_scan_restrictions (struct scanner *, int);
+init_scan_restrictions (scanner_t *, int);
 
 void
-handle_scan_restrictions (struct scanner *, gchar *);
+handle_scan_restrictions (scanner_t *, gchar *);
 
 gchar *
 get_openvas_scan_id (const gchar *, int);
@@ -55,5 +55,8 @@ get_alive_test_methods (alive_test_t *);
 
 const gchar *
 get_alive_test_ports (void);
+
+int
+get_alive_hosts_count (void);
 
 #endif /* not BOREAS_IO_H */