+++ boreas/cli.c
@@ -20,6 +20,7 @@
 #include "cli.h"
 
 #include "../base/networking.h"
+#include "../base/prefs.h"
 #include "alivedetection.h"
 #include "boreas_io.h"
 #include "ping.h"
@@ -34,11 +35,11 @@
 /**
  * @brief GLib log domain.
  */
-#define G_LOG_DOMAIN "alive scan"
+#define G_LOG_DOMAIN "libgvm boreas"
 
 static boreas_error_t
-init_cli (struct scanner *scanner, gvm_hosts_t *hosts, alive_test_t alive_test,
-          const gchar *port_list)
+init_cli (scanner_t *scanner, gvm_hosts_t *hosts, alive_test_t alive_test,
+          const gchar *port_list, const int print_results)
 {
   GPtrArray *portranges_array;
   gvm_host_t *host;
@@ -48,6 +49,7 @@ init_cli (struct scanner *scanner, gvm_hosts_t *hosts, alive_test_t alive_test,
 
   /* No kb used for cli mode.*/
   scanner->main_kb = NULL;
+  scanner->print_results = print_results;
 
   /* hosts_data */
   scanner->hosts_data = g_malloc0 (sizeof (hosts_data_t));
@@ -81,7 +83,7 @@ init_cli (struct scanner *scanner, gvm_hosts_t *hosts, alive_test_t alive_test,
 }
 
 static boreas_error_t
-free_cli (struct scanner *scanner, alive_test_t alive_test)
+free_cli (scanner_t *scanner, alive_test_t alive_test)
 {
   int close_err;
 
@@ -99,7 +101,7 @@ free_cli (struct scanner *scanner, alive_test_t alive_test)
 }
 
 static boreas_error_t
-run_cli_scan (struct scanner *scanner, alive_test_t alive_test)
+run_cli_scan (scanner_t *scanner, alive_test_t alive_test)
 {
   int error;
   int number_of_dead_hosts;
@@ -110,7 +112,8 @@ run_cli_scan (struct scanner *scanner, alive_test_t alive_test)
   gettimeofday (&start_time, NULL);
   number_of_targets = g_hash_table_size (scanner->hosts_data->targethosts);
 
-  printf ("Alive scan started: Target has %d hosts.\n", number_of_targets);
+  if (scanner->print_results == 1)
+    printf ("Alive scan started: Target has %d hosts.\n", number_of_targets);
 
   error = start_sniffer_thread (scanner, &sniffer_thread_id);
   if (error)
@@ -158,9 +161,10 @@ run_cli_scan (struct scanner *scanner, alive_test_t alive_test)
   number_of_dead_hosts = count_difference (scanner->hosts_data->targethosts,
                                            scanner->hosts_data->alivehosts);
   gettimeofday (&end_time, NULL);
-  printf ("Alive scan finished in %ld seconds: %d alive hosts of %d.\n",
-          end_time.tv_sec - start_time.tv_sec,
-          number_of_targets - number_of_dead_hosts, number_of_targets);
+  if (scanner->print_results == 1)
+    printf ("Alive scan finished in %ld seconds: %d alive hosts of %d.\n",
+            end_time.tv_sec - start_time.tv_sec,
+            number_of_targets - number_of_dead_hosts, number_of_targets);
 
   return error;
 }
@@ -168,12 +172,13 @@ run_cli_scan (struct scanner *scanner, alive_test_t alive_test)
 boreas_error_t
 run_cli (gvm_hosts_t *hosts, alive_test_t alive_test, const gchar *port_list)
 {
-  struct scanner scanner = {0};
+  scanner_t scanner = {0};
   boreas_error_t init_err;
   boreas_error_t run_err;
   boreas_error_t free_err;
+  int print_results = 1;
 
-  init_err = init_cli (&scanner, hosts, alive_test, port_list);
+  init_err = init_cli (&scanner, hosts, alive_test, port_list, print_results);
   if (init_err)
     {
       printf ("Error initializing scanner.\n");
@@ -186,6 +191,64 @@ run_cli (gvm_hosts_t *hosts, alive_test_t alive_test, const gchar *port_list)
       printf ("Error while running the scan.\n");
       return run_err;
     }
+
+  free_err = free_cli (&scanner, alive_test);
+  if (free_err)
+    {
+      printf ("Error freeing scan data.\n");
+      return free_err;
+    }
+
+  return NO_ERROR;
+}
+
+/**
+ * @brief Scan all specified hosts in ip_str list.
+ *
+ * @param[in] ip_str list of hosts to alive test.
+ * @param[out] count amount of alive host which were found alived.
+ *
+ * @return NO_ERROR (0) on success, boreas_error_t on error.
+ */
+boreas_error_t
+is_host_alive (const char *ip_str, int *count)
+{
+  scanner_t scanner = {0};
+  boreas_error_t init_err;
+  boreas_error_t run_err;
+  boreas_error_t free_err;
+  boreas_error_t alive_test_err;
+  alive_test_t alive_test;
+  gvm_hosts_t *hosts;
+  const int print_results = 0;
+  const gchar *port_list = NULL;
+
+  hosts = gvm_hosts_new (ip_str);
+  if ((alive_test_err = get_alive_test_methods (&alive_test)) != 0)
+    {
+      g_warning ("%s: %s. Exit Boreas.", __func__,
+                 str_boreas_error (alive_test_err));
+      pthread_exit (0);
+    }
+
+  /* This port list is also used by openvas for scanning and was already
+   * validated by openvas so we don't do it here again. */
+  port_list = prefs_get ("port_range");
+
+  init_err = init_cli (&scanner, hosts, alive_test, port_list, print_results);
+  if (init_err)
+    {
+      printf ("Error initializing scanner.\n");
+      return init_err;
+    }
+
+  run_err = run_cli_scan (&scanner, alive_test);
+  if (run_err)
+    {
+      printf ("Error while running the scan.\n");
+      return run_err;
+    }
+  *count = get_alive_hosts_count ();
 
   free_err = free_cli (&scanner, alive_test);
   if (free_err)