+++ boreas/cli.h
@@ -26,4 +26,7 @@
 boreas_error_t
 run_cli (gvm_hosts_t *, alive_test_t, const gchar *);
 
+boreas_error_t
+is_host_alive (const char *, int *);
+
 #endif /* not CLI_H */