+++ util/kb.c
@@ -35,7 +35,10 @@
 #include <string.h> /* for strlen, strerror, strncpy, memset */
 
 #undef G_LOG_DOMAIN
-#define G_LOG_DOMAIN "lib  kb"
+/**
+ * @brief GLib logging domain.
+ */
+#define G_LOG_DOMAIN "libgvm util"
 
 /**
  * @file kb.c
@@ -276,7 +279,6 @@ get_redis_ctx (struct kb_redis *kbr)
              kbr->rctx ? kbr->rctx->errstr : strerror (ENOMEM));
       redisFree (kbr->rctx);
       kbr->rctx = NULL;
-      g_free (kbr->path);
       return -1;
     }
 
@@ -286,7 +288,6 @@ get_redis_ctx (struct kb_redis *kbr)
       g_log (G_LOG_DOMAIN, G_LOG_LEVEL_CRITICAL, "No redis DB available");
       redisFree (kbr->rctx);
       kbr->rctx = NULL;
-      g_free (kbr->path);
       return -2;
     }
 
@@ -446,7 +447,7 @@ redis_new (kb_t *kb, const char *kb_path)
 
   /* Try to make unused memory available for the OS again. */
   if (redis_memory_purge (*kb))
-    g_warning ("%s: Memory purge was not successfull", __func__);
+    g_warning ("%s: Memory purge was not successful", __func__);
 
   return rc;
 }